# open-api/openapi.yaml
openapi: 3.0.3
info:
  title: TELETUBBIES COURSE API
  description: |
    This API is focused on performing CRUD type operations over courses and instructors.
  contact:
    email: contacto@iwa.com.mx
  version: 0.0.1
  x-logo:
    backgroundColor: '#FFFFFF'
    altText: 'Teletubbies Course'

servers:
  - url: 'http://{baseUrl}:{port}'
    description: local
    variables:
      port:
        enum:
          - '8080'
        default: '8080'
      baseUrl:
        default: localhost

tags:
  - name: courses
  - name: instructors

paths:
  '/courses':
    post:
      tags:
        - courses
      summary: Add a new course
      description: |
        This endpoint allows you to create a new course using a form.
      operationId: createCourse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'requests.yaml#/components/schemas/NewCourseRequest'
      responses:
        201:
          $ref: 'responses.yaml#/components/responses/CourseResponse'
        400:
          $ref: 'responses.yaml#/components/responses/BadRequest'
        401:
          $ref: 'responses.yaml#/components/responses/Unauthorized'
        403:
          $ref: 'responses.yaml#/components/responses/Forbidden'
        404:
          $ref: 'responses.yaml#/components/responses/NotFound'
        405:
          $ref: 'responses.yaml#/components/responses/MethodNotAllowed'
        409:
          $ref: 'responses.yaml#/components/responses/Conflict'
        500:
          $ref: 'responses.yaml#/components/responses/InternalServerError'

    get:
      tags:
        - courses
      summary: Get all courses
      description: |
        This endpoint retrieves a paginated list of courses with optional filters by name, level,
        category and instructor.
      operationId: getAllCourses
      parameters:
        - name: pageable
          in: query
          required: false
          schema:
            $ref: './parameters.yaml#/components/parameters/Pageable'
        - $ref: 'parameters.yaml#/components/parameters/CourseNameFilter'
        - $ref: 'parameters.yaml#/components/parameters/CourseLevelFilter'
        - $ref: 'parameters.yaml#/components/parameters/CourseCategoryFilter'
        - $ref: 'parameters.yaml#/components/parameters/CourseInstructorIdFilter'
      responses:
        200:
          $ref: 'responses.yaml#/components/responses/CoursesPageResponse'
        401:
          $ref: 'responses.yaml#/components/responses/Unauthorized'
        403:
          $ref: 'responses.yaml#/components/responses/Forbidden'
        405:
          $ref: 'responses.yaml#/components/responses/MethodNotAllowed'
        500:
          $ref: 'responses.yaml#/components/responses/InternalServerError'

  '/courses/{courseId}':
    put:
      tags:
        - courses
      summary: Update course
      description: |
        This endpoint allows you to update an existing course based on its ID using a form.
      operationId: updateCourse
      parameters:
        - $ref: 'parameters.yaml#/components/parameters/CourseId'
      requestBody:
        description: CourseForm contains properties of the course to be updated.
        required: true
        content:
          application/json:
            schema:
              $ref: 'requests.yaml#/components/schemas/UpdateCourseRequest'
      responses:
        200:
          $ref: 'responses.yaml#/components/responses/CourseResponse'
        400:
          $ref: 'responses.yaml#/components/responses/BadRequest'
        401:
          $ref: 'responses.yaml#/components/responses/Unauthorized'
        403:
          $ref: 'responses.yaml#/components/responses/Forbidden'
        404:
          $ref: 'responses.yaml#/components/responses/NotFound'
        405:
          $ref: 'responses.yaml#/components/responses/MethodNotAllowed'
        409:
          $ref: 'responses.yaml#/components/responses/Conflict'
        500:
          $ref: 'responses.yaml#/components/responses/InternalServerError'

    delete:
      tags:
        - courses
      summary: Delete course
      description: |
        This endpoint allows deleting an existing course based on its ID.
      operationId: deleteCourse
      parameters:
        - $ref: 'parameters.yaml#/components/parameters/CourseId'
      responses:
        204:
          $ref: 'responses.yaml#/components/responses/Successful'
        401:
          $ref: 'responses.yaml#/components/responses/Unauthorized'
        403:
          $ref: 'responses.yaml#/components/responses/Forbidden'
        404:
          $ref: 'responses.yaml#/components/responses/NotFound'
        405:
          $ref: 'responses.yaml#/components/responses/MethodNotAllowed'
        500:
          $ref: 'responses.yaml#/components/responses/InternalServerError'
    get:
      tags:
        - courses
      summary:  Get individual course
      description: |
        This endpoint return an existing course based on its ID.
      operationId: getCourse
      parameters:
        - $ref: 'parameters.yaml#/components/parameters/CourseId'
      responses:
        200:
          $ref: 'responses.yaml#/components/responses/CourseResponse'
        401:
          $ref: 'responses.yaml#/components/responses/Unauthorized'
        403:
          $ref: 'responses.yaml#/components/responses/Forbidden'
        404:
          $ref: 'responses.yaml#/components/responses/NotFound'
        405:
          $ref: 'responses.yaml#/components/responses/MethodNotAllowed'
        500:
          $ref: 'responses.yaml#/components/responses/InternalServerError'

  '/instructors':
    post:
      tags:
        - instructors
      summary: Add a new instructor
      description: |
        This endpoint allows you to create a new instructor using a form.
        Full name and email must be unique.
      operationId: createInstructor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'requests.yaml#/components/schemas/NewInstructorRequest'
      responses:
        201:
          $ref: 'responses.yaml#/components/responses/InstructorResponse'
        400:
          $ref: 'responses.yaml#/components/responses/BadRequest'
        401:
          $ref: 'responses.yaml#/components/responses/Unauthorized'
        403:
          $ref: 'responses.yaml#/components/responses/Forbidden'
        405:
          $ref: 'responses.yaml#/components/responses/MethodNotAllowed'
        409:
          $ref: 'responses.yaml#/components/responses/Conflict'
        500:
          $ref: 'responses.yaml#/components/responses/InternalServerError'

    get:
      tags:
        - instructors
      summary: Get all instructors
      description: |
        This endpoint retrieves a paginated list of instructors with optional filters by full
        name, email and role.
      operationId: getAllInstructors
      parameters:
        - name: pageable
          in: query
          required: false
          schema:
            $ref: './parameters.yaml#/components/parameters/Pageable'
        - $ref: 'parameters.yaml#/components/parameters/InstructorFullNameFilter'
        - $ref: 'parameters.yaml#/components/parameters/InstructorEmailFilter'
        - $ref: 'parameters.yaml#/components/parameters/InstructorRoleFilter'
      responses:
        200:
          $ref: 'responses.yaml#/components/responses/InstructorsPageResponse'
        401:
          $ref: 'responses.yaml#/components/responses/Unauthorized'
        403:
          $ref: 'responses.yaml#/components/responses/Forbidden'
        405:
          $ref: 'responses.yaml#/components/responses/MethodNotAllowed'
        500:
          $ref: 'responses.yaml#/components/responses/InternalServerError'

  '/instructors/{instructorId}':
    get:
      tags:
        - instructors
      summary: Get individual instructor
      description: |
        This endpoint returns an existing instructor based on its ID.
      operationId: getInstructor
      parameters:
        - $ref: 'parameters.yaml#/components/parameters/InstructorId'
      responses:
        200:
          $ref: 'responses.yaml#/components/responses/InstructorResponse'
        401:
          $ref: 'responses.yaml#/components/responses/Unauthorized'
        403:
          $ref: 'responses.yaml#/components/responses/Forbidden'
        404:
          $ref: 'responses.yaml#/components/responses/NotFound'
        405:
          $ref: 'responses.yaml#/components/responses/MethodNotAllowed'
        500:
          $ref: 'responses.yaml#/components/responses/InternalServerError'

    put:
      tags:
        - instructors
      summary: Update instructor
      description: |
        This endpoint allows you to update an existing instructor based on its ID using a form.
        Full name and email must remain unique.
      operationId: updateInstructor
      parameters:
        - $ref: 'parameters.yaml#/components/parameters/InstructorId'
      requestBody:
        description: InstructorForm contains properties of the instructor to be updated.
        required: true
        content:
          application/json:
            schema:
              $ref: 'requests.yaml#/components/schemas/UpdateInstructorRequest'
      responses:
        200:
          $ref: 'responses.yaml#/components/responses/InstructorResponse'
        400:
          $ref: 'responses.yaml#/components/responses/BadRequest'
        401:
          $ref: 'responses.yaml#/components/responses/Unauthorized'
        403:
          $ref: 'responses.yaml#/components/responses/Forbidden'
        404:
          $ref: 'responses.yaml#/components/responses/NotFound'
        405:
          $ref: 'responses.yaml#/components/responses/MethodNotAllowed'
        409:
          $ref: 'responses.yaml#/components/responses/Conflict'
        500:
          $ref: 'responses.yaml#/components/responses/InternalServerError'

    delete:
      tags:
        - instructors
      summary: Delete instructor
      description: |
        This endpoint allows deleting an existing instructor based on its ID.
        An instructor assigned to courses cannot be deleted.
      operationId: deleteInstructor
      parameters:
        - $ref: 'parameters.yaml#/components/parameters/InstructorId'
      responses:
        204:
          $ref: 'responses.yaml#/components/responses/Successful'
        401:
          $ref: 'responses.yaml#/components/responses/Unauthorized'
        403:
          $ref: 'responses.yaml#/components/responses/Forbidden'
        404:
          $ref: 'responses.yaml#/components/responses/NotFound'
        405:
          $ref: 'responses.yaml#/components/responses/MethodNotAllowed'
        409:
          $ref: 'responses.yaml#/components/responses/Conflict'
        500:
          $ref: 'responses.yaml#/components/responses/InternalServerError'